requirements.txt
alembic==0.9.1
amqp==2.1.4
appdirs==1.4.3
appnope==0.1.0
billiard==3.5.0.2
celery==4.0.2
click==6.7
-e git+git@github.com:getweber/cob.git@d9ce1f7039458566d0a4614f6630bccca719c452#egg=cob
colorama==0.3.7
decorator==4.0.11
emport==1.2.0
Flask==0.12
Flask-Loopback==1.4.5
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
gossip==2.1.1
gunicorn==19.7.0
ipython==5.3.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.9.5
kaptan==0.5.8
kombu==4.0.2
libtmux==0.6.3
Logbook==1.0.0
Mako==1.0.6
MarkupSafe==0.23
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
psycopg2==2.7.1
ptyprocess==0.5.1
pudb==2017.1.1
py==1.4.32
Pygments==2.2.0
pyldap==2.4.28
pyparsing==2.2.0
pytest==3.0.6
python-editor==1.0.3
pytz==2016.10
PyYAML==3.12
requests==2.12.4
sentinels==1.0.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.6
SQLAlchemy-Utils==0.32.14
tmuxp==1.2.6
traitlets==4.3.2
URLObject==2.4.2
urwid==1.3.1
vine==1.1.3
vintage==0.3.0
virtualenv==15.1.0
wcwidth==0.1.7
Werkzeug==0.12
